refactor: 라우팅 없이 결제 위젯 연결
