+++ src/pages/PaymentsCheckout.tsx
@@ -1,142 +1,32 @@
-import { useEffect, useState } from 'react';
-import { loadTossPayments } from '@tosspayments/tosspayments-sdk';
-import { nanoid } from 'nanoid';
+import { useFunnel } from '@/hooks/common/useFunnel';
+import { Payments } from '@/components/payments/Payments';
+import { PaymentsWidget } from '@/components/payments/PaymentsWidget';
+import useCustomBack from '@/hooks/common/useCutomBack';
 
-import { useQuery } from '@tanstack/react-query';
-import RoutePath from '@/routes/routePath';
+const steps = ['회비 납부', '결제 위젯'];
 
-import { media } from '@/styles';
-import { Flex, Space } from '@/components/common/Wrapper';
-import Button from 'wowds-ui/Button';
+export const PaymentsCheckout = () => {
+  const { Funnel, Step, setStep, currentStep } = useFunnel(steps[0]);
 
-import GlobalSize from '@/constants/globalSize';
-import styled from '@emotion/styled';
-import { CLIENT_KEY } from '@/constants/environment';
-import { color } from 'wowds-tokens';
-import meApi from '@/apis/me/meApi';
-import ordersApi from '@/apis/orders/ordersApi';
-import memberApi from '@/apis/member/memberApi';
-import { useProduct } from '@/hooks/zustand/useProduct';
-
-const clientKey = CLIENT_KEY;
-const customerKey = nanoid();
-
-export function PaymentsCheckout() {
-  const { data: user } = useQuery({
-    queryKey: ['me'],
-    queryFn: meApi.GET_BASIC_INFO
-  });
-
-  const { data: dashboard } = useQuery({
-    queryKey: ['member'],
-    queryFn: memberApi.GET_DASHBOARD
-  });
-
-  const { name, amount, discount, issuedCouponId, totalAmount } = useProduct();
-
-  const [ready, setReady] = useState(false);
-
-  // 토스페이먼츠에서 타입을 제공해주지 않아 임시로 타입을 any로 설정합니다.
-  // eslint-disable-next-line
-  const [widgets, setWidgets] = useState<any>(null);
-
-  useEffect(() => {
-    async function fetchPaymentWidgets() {
-      try {
-        const tossPayments = await loadTossPayments(clientKey);
-        const widgets = tossPayments.widgets({
-          customerKey
-        });
-
-        setWidgets(widgets);
-      } catch (error) {
-        console.error('Error fetching payment widget:', error);
-      }
-    }
-
-    fetchPaymentWidgets();
-  }, [clientKey, customerKey]);
-
-  useEffect(() => {
-    async function renderPaymentWidgets() {
-      if (widgets == null) {
-        return;
-      }
-      await widgets.setAmount({ currency: 'KRW', value: totalAmount });
-      await widgets.renderPaymentMethods({
-        selector: '#payment-method',
-        variantKey: 'DEFAULT'
-      });
-
-      await widgets.renderAgreement({
-        selector: '#agreement',
-        variantKey: 'AGREEMENT'
-      });
-
-      setReady(true);
-    }
-
-    renderPaymentWidgets();
-  }, [widgets]);
-
-  const handleClickOpenPaymentWidget = async () => {
-    const id = nanoid();
-    try {
-      if (!dashboard) throw new Error();
-
-      await ordersApi.POST_PREV_ORDER({
-        orderNanoId: id,
-        membershipId: dashboard.currentMembership.membershipId,
-        issuedCouponId: issuedCouponId,
-        totalAmount: amount,
-        discountAmount: discount,
-        finalPaymentAmount: totalAmount
-      });
-
-      await widgets.requestPayment({
-        orderId: id,
-        orderName: name,
-        customerName: user?.name,
-        customerEmail: user?.email,
-        customerMobilePhone: user?.phone.replaceAll('-', ''),
-        successUrl: `${window.location.origin}${RoutePath.PaymentsSuccess}`,
-        failUrl: `${window.location.origin}${RoutePath.PaymentsFail}`
-      });
-    } catch (error) {
-      console.error(error);
-    }
+  const nextClickHandler = (step: string) => {
+    setStep(step);
   };
 
+  const handleBack = () => {
+    const currentStepIndex = steps.indexOf(currentStep);
+    if (currentStepIndex === 0) return;
+    setStep(steps[currentStepIndex - 1]);
+  };
+
+  useCustomBack(handleBack);
   return (
-    <Wrapper direction="column" justify="space-between">
-      <Contents className="box_section">
-        <div id="payment-method" />
-        <div id="agreement" />
-      </Contents>
-      <Flex direction="column">
-        <Button disabled={!ready} onClick={handleClickOpenPaymentWidget}>
-          결제하기
-        </Button>
-        <Space height={28} />
-      </Flex>
-    </Wrapper>
+    <Funnel>
+      <Step name={steps[0]}>
+        <Payments onNext={() => nextClickHandler(steps[1])} />
+      </Step>
+      <Step name={steps[1]}>
+        <PaymentsWidget />
+      </Step>
+    </Funnel>
   );
-}
-
-const Wrapper = styled(Flex)`
-  box-sizing: border-box;
-
-  height: calc(100vh - ${GlobalSize.header});
-  width: ${GlobalSize.width};
-  padding: 0px 16px;
-
-  background-color: ${color.backgroundAlternative};
-
-  ${media.mobile} {
-    width: 100vw;
-  }
-`;
-
-const Contents = styled.div`
-  width: 100%;
-`;
+};
