+++ src/pages/index.ts
@@ -10,7 +10,6 @@ export { default as OnboardingNotOpened } from './OnboardingNotOpened';
 export { default as OnboardingClosed } from './OnboardingClosed';
 export * from './redirect/AuthServerRedirectNavigate';
 export * from './redirect/StudentVerificationServerRedirect';
-export * from './Payments';
 export * from './PaymentsCheckout';
 export * from './PaymentsFail';
 export * from './PaymentsSuccess';
